requirements.txt
pypdf
python-docx
pdfplumber

google-generativeai
openai
python-dotenv

fastapi
uvicorn
pydantic
python-multipart

SQLAlchemy
alembic
psycopg2-binary
apache-airflow-providers-postgres

pytest